Completed Wiremock setup
