+++ web-backend/src/main/java/com/statusneo/vms/service/GraphEmailService.java
@@ -2,6 +2,7 @@ package com.statusneo.vms.service;
 
 import com.statusneo.vms.model.Email;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
 import org.springframework.http.*;
 import org.springframework.security.oauth2.client.OAuth2AuthorizeRequest;
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClient;
@@ -19,6 +20,7 @@ import java.util.Map;
  * Manages visitor notifications, OTP delivery, and employee communications.
  */
 @Service
+@Profile("prod")
 public class GraphEmailService implements EmailService {
 
     private final OAuth2AuthorizedClientManager authorizedClientManager;
