+++ web-backend/src/main/java/com/statusneo/vms/service/WiremockMailServiceImpl.java
@@ -20,9 +20,6 @@ public class WiremockMailServiceImpl implements EmailService{
 
     private final RestTemplate restTemplate;
 
-    @Value("${msgraph.base-url}")
-    private String graphBaseUrl;
-
     public WiremockMailServiceImpl(RestTemplate restTemplate) {
         this.restTemplate = restTemplate;
     }
@@ -31,7 +28,7 @@ public class WiremockMailServiceImpl implements EmailService{
      *
      * */
     public boolean sendEmail(Email email) {
-        String endpoint = String.format(graphBaseUrl + "/v1.0/%s/sendMail", email.from());
+        String endpoint = String.format("http://localhost:8089/v1.0/users/%s/sendMail", email.from());
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
